bootmenu: Avoid building unless it is required.
I know this breaks the standalone build, this is solely for the
purpose of commenting towards a solution.

Not all devices will produce a UI library if they do not produce
a recovery. This is the case for the generic target which is
unbuildable with the current state of bootmenu.

Change-Id: Ib6338c3c1b9f79b01e8c9ade8b2f72760c26fc7a